requirements.txt
beautifulsoup4
gTTS
googlesearch-python
playsound>=1.2.2
requests
